+++ dev/Examples.vue
@@ -31,6 +31,7 @@ import Description from '../components/Description.vue';
 import ExperimentalNotice from '../components/ExperimentalNotice.vue';
 import FileFormat from '../components/FileFormat.vue';
 import FileFormats from '../components/FileFormats.vue';
+import Item from '../components/Item.vue';
 import JsonSchema from '../components/JsonSchema.vue';
 import LinkList from '../components/LinkList.vue';
 import ObjectTree from '../components/ObjectTree.vue';
@@ -57,6 +58,7 @@ export default {
 		ExperimentalNotice,
 		FileFormat,
 		FileFormats,
+		Item,
 		JsonSchema,
 		LinkList,
 		ObjectTree,